Moved plotting routines to separate submodule
